Optionally support no_std (#31)
Add a "std" feature which is enabled by default, and enable `no_std`
when it is not enabled. Disable all `Display` and `Debug` features
in `no_std` mode.

My specific use case for this is [rsix](https://crates.io/crates/rsix),
which has a no_std branch, and which uses the `errno` crate in its
`libc`-using backend.

Fixes #18.