Switch away from actions-rs
The actions-rs organization is unmaintained with the sole member being
unresponsive for two years.

The actions still use Node 12 and the `set-output` command, both of which are deprecated and will
be removed by Github Actions by Summer 2023. [1][2]

Switch from `actions-rs/toolchain` to `dtolnay/rust-toolchain` and from `actions-rs/cargo` to plain
cargo commands.

[1] https://github.blog/changelog/2022-09-22-github-actions-all-actions-will-begin-running-on-node16-instead-of-node12/
[2] https://github.blog/changelog/2022-10-11-github-actions-deprecating-save-state-and-set-output-commands/