Add `rust-version` field (#62)
This field tells cargo about errno's MSRV. It was added with Rust 1.56. Older versions will print a
warning about an unknown field if errno is the top-level build job. If errno is merely compiled as
a dependency, no warning is printed.

Even though errno's MSRV is currenlty 1.48, it would be good to make this metadata available in a
standarized machine-readable format.