wasm32-wasi: Use `__errno_location` instead of `feature(thread_local)`. (#66)
Recent versions of wasi-libc provide an `__errno_location` function
which returns the address of errno, similar to other platforms. Change
the errno crate to use that, instead of using the unstable
`feature(thread_local)`.

At the moment, only Rust nightly has a new enough wasi-libc to support
this, however it's not a regression because the `feature(thread_local)`
already dependend on nightly. In the future, the newer wasi-libc will
be in stable Rust too, making this work on stable.